requirements.txt
altair==5.0.1
beautifulsoup4==4.12.3
matplotlib==3.7.1
numpy==1.25.0
Pillow==9.5.0
pandas==2.0.2
streamlit==1.23.1